user_watch_management/requirements.txt
Django==4.1.4
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.2.2
celery==5.2.7
redis==4.4.2
Pillow==9.4.0
